Change behavior when empty custom amount is entered
Use case:
One of the preset amounts is selected.
The custom amount field is clicked.
No value is entered in the custom amount field.
The focus leaves the custom amount field.

Don't clear the selected amount and don't show an error message if the custom amount field is on focus but no value is entered.