Add test for payment interval component
Refactor view model and component to use simple month durations as intervals.

Add donation intervals to presenter to keep server and client side in
sync. This is by no means extensive (Intervals and the collection of
possible intervals should be their own domain objects) and still only
used in the laika skin, but it's a step towards a more expressive
domain.