Refactor DonationSummary to standalone component
Move some of the layout to DonationConfirmation
Move links and link logic to SummaryLinks component
Move Payment summary text to PaymentSummary component.
Use div instead of span with display:block, as the entries are also more
semantic "blocks" than "inline elements".
Add `canRender` method to display error message if not all address
fields are filled (This will be used in the donation form)

This refactoring is in preparation for re-using DonationSummary at the
bottom of the donation page.