Split DonationForm into pages
Since the button areas for payment and address pages are radically
different, it's more convenient to create pages for them that each do the
previous/next error logic and relegate the DonationForm to a simple page
switcher.

Removed some tests because now the "behavior" is implicit in the page
markup.

Fixed amount preview in the summary.