Refactor Payment Type
While the payment type won't be reused currently on the membership page,
the pattern it follows should be the same as the interval selection,
decoupled from the store. This makes the code more consistent and ready
for introducing different payment types on the membership page.

Remove dependency on store from the shared component. This is also a
better solution for the Buefy radio component needing a model.
Add event handling for payment type selection.
Adapt unit tests.