Extend country input field functionality
The DonationCountry.ts interface might be useless,
because it is only needed for client side validation.
Client-side validation is not needed as acceptance criteria in the ticket.

The country input field still needs a working default choice.

The error message for validation might have to be adjusted.

The list of countries has to be updated and centralized.

The functionality still has to be tested, and implemented for memebership and address update.

https://phabricator.wikimedia.org/T247227