Remove translations from presentation layer
Remove the Translator instance from Twig initialization and the `trans`
filter from the twig templates.

All translated values that have to be rendered in HTML generated by the
PHP side instead of the client-side are stored in the new
`siteMetadata.json` file, assigned to all templates.

Change tests that access content on the confirmation page to use
assigned values instead of rendered output.