Add CI job
Add a CI job that runs linting and unit testing tasks.
Add caching to the tasks to avoid unnecessary downloads of npm modules
(cache key is using the branch name, so there will be downloads of
modules in new branches).

Move docker image build jobs to "package" stage - the docker images are
not the primary "build outputs" of this repo, they are artifacts.

Remove "build" stage - we don't have a compiled language here, no need
to build anything.