Re-Activate fonts
Remove commented-out fonts from custom.scss

Add webpack configuration to leave `url` statements on CSS alone.

Add webpack plugin to copy the `public` directory into `dist`.

If there are issues with missing fonts when using the dev server (will
be activated in one of the upcoming commits) the refer to
https://github.com/webpack-contrib/copy-webpack-plugin/issues/29