Use null object as default
To make Typescript happy, we inject a NullObject implementation of the
CookieConsent interface as default, so we no longer need a `null` check
that needs to be tested (and can't be tested properly).

With the current implementation, the developer is no longer warned when
they forget to inject the real implementation.