Make 2-step address type default
Remove the A/B test of the two address types, making the 2-step address
form (2nd step is choosing person/company) the default.

Adapt unit tests for AddressType and Address.
The tests that check the interaction between form fields and the store
are skipped for now because the new address component does not display
any input fields and it's hard to display them (to trigger store changes).