Convert name component to composition API
This avoids having to expose AddressTypeModel as a computed property and
allows for more semantic computed properties.

Removed the "prevent placeholder as last name" feature because I've
discovered that it is broken and won't work with the current placeholder
anyway. We need to get input in how to fix it, tracked in
https://phabricator.wikimedia.org/T286540

Anticipating the "contact person for company" task, I've also changed an
else-if into an if. See https://phabricator.wikimedia.org/T220366