Revalidate fee on membership type change
The server applies different minimum membership fees for people and
companies. Therefore, we trigger validation whenever the membership
address type changes.

Triggering from the Vue component has benefits and downsides:

Benefit: the separation between address store and fee store is still
intact.

Drawback: The separation of concerns (state and UI) is muddled and the
whole system becomes more complicated. Also, the behavior is hard to
test and currently untested (last-day-of-campaign pressure).

This commit further entrenches the technical debt that is the state
management - not splitting out validation as a service (see all the
passing around of validation URLs through 3 components), and separating
state that needs to be validated together, creating fake boundaries that
are hard to test.

This fixes https://phabricator.wikimedia.org/T298368