Update ESLint Vue and TS dependencies
Update @vue/eslint-config-typescript and eslint-plugin-vue
(eslint-config-typescript version 10 depends on eslint-plugin-vue 8)

Downgrade TypeScript to 4.4 because eslint-config-typescript does not support
TypeScript 4.5 right now.

vue-eslint-config-typescript-10 expects the plugins to be defined in a
separate package and using @rushstack/eslint-patch to patch the eslint
config to pick up the plugin. We're not doing that, instead we just
install the missing library