Update to Jest 27
Update Jest to Version 27
Change vue-jest dependency to @vue/vue2-jest
Adapt jest config

Jest 27 changes the default test environment from "jsdom" to "node".
This means that tests which can run without a DOM environment will be
faster, since they don't need to initialize jsdom.
Most of our tests are component tests, so we're setting "jsdom" as the
default.

In the future, we should figure out how to run store and other unit
tests in the node environment. This could be done with a
[testEnvironment](https://jestjs.io/docs/configuration#testenvironment-string)
annotation in the class files or by configuring different
[projects](https://jestjs.io/docs/configuration#projects-arraystring--projectconfig)

This is for https://phabricator.wikimedia.org/T299902