Hide payment selection depending on availability
When only 1 payment is available for memberships, (configured in the
application), pre-select that payment and hide the payment selection

This also fixes a TypeScript "error" where TypeScrupt does not recognize
our call to Vuex "mapGetters" and complains about an unknown getter. The
workaround is to access the getter directly.

This is for https://phabricator.wikimedia.org/T312072