Catch hitting Enter in custom amount
In some cases hitting Enter would submit the form, bypassing the
validation and setting of values, leading to validation errors.

This new behavior makes the usability a bit worse, since it breaks
default browser functionality. However, a change to preserve it would be
quite complicated because of our strict separation of concerns between
components (sending data from events to the store and rendering values
from the store) and store (keeping the state and validating it, sending
validation to the sever and processing it). To preserve the Enter key
functionality, we'd need to check if only one field is displayed,
trigger a second event and make the event handler for that event install
a temporary listener for store state changes to be notified then the
validation has finished and all fields are valid.

Ticket: https://phabricator.wikimedia.org/T322393