Change transl keys for existing AB test component
- we currently have already 3 different Address Type options
in a feature test:
 - AddressTypeAllOptions.vue
 - AddressTypeFullOrEmail.vue
 - AddressTypeBasic.vue

for this ticket, only the translation keys of the existing FullOrEmail component
needed to be changed. The options now ask if a user
wants a donation receipt or just a confirmation mail.

The change can be accessed via ?ast=2 as a URL parameter in the browser.
Further AB test settings can befound in fundraising-application.

- this also removes the little info text from this component

https://phabricator.wikimedia.org/T342206