Fix issues with application vars
Use of Funds, FAQ, and supporters pages were not
outputting the application vars. This was fixed in
the fundraising application repository, but now
the data on those pages needs to be parsed into json.

This also adds the bucket classes to every entry point.

Ticket: https://phabricator.wikimedia.org/T345721