Add donation survey tile to donation confirmation page.
The donation survey tile asks users to participate in a short survey.
The link in the tile leads to either the German or the English version of the survey.
Copy and link targets are in the content repository.

Change snake case to Pascal case for using Vue components in template

Ticket: https://phabricator.wikimedia.org/T349193