Add a sample donation form accordion
We use a fake 2 page form for donations. The problem
with that is the donors may expect the browser back
button to bring them to the first page. Hijacking the
browser history to try manage this causes issues so
this is a test of a single page form.

Ticket: https://phabricator.wikimedia.org/T339109