Fix some eslint a11y issues
The ESLint plugin `eslint-plugin-vuejs-accessibility` marks some
accessibility issues. This commit tackles the issues that won't be
tackled by other, dedicated tickets and pull requests.

We're mostly ignoring the "label" rule inside the files. The files where
we disable it are "false positives" where we're using our own input
components (that aren't recognized as inputs by the rule). In the
future, we might unify the form field components to avoid the A11Y issues
and ensure that IDs and labels do match.

We default on "Enter" and "Space" as the keyboard alternative to "click"
events on non-interactive elements that have a click handler.

We have to disable the `no-static-element-interactions` rule on the
`PaymentForm` component because its keyboard event handler is supposed
to catch the Enter key that bubbles up from other elements.