Add DirectDebitField component
This is drop in replacement for the PaymentBankData component
that is more accessible and integrated with our form fields
components.

- Added accountNumber and bankCode fields, as the direct debit
  field will accept both IBAN and account numbers making the
  naming confusing.
- Added getters actions and mutations for the new fields.
- Updated the typescript interfaces to reflect the new structure.

This makes the API calls that validate the bank data more
consistent with our other API resources

The pattern for our form fields is that they take in model
values and do not access the store directly, We then use
wrapper components to manage store access. That's what this
component is for.

This handles visibility of labels, and formatting values
when fields receive input.

As we're formatting the value in the input box we need
a field similar to TextField that uses a prop as the value
rather than take in a model.

Ticket: https://phabricator.wikimedia.org/T364953