Fix missing prop in city autocomplete
This form field needs to know the scroll target so it
can scroll the element into view when focused and make
the autocomplete list visible.

It was set up to do the scroll but in some of the places
the component was being used it wasn't being passed the
scroll target id.