Add street autocomplete to alternate address form
We're about to do a test that needs both features
so now both address forms have both form orders.

The bank validation error in the summary wasn't appearing
so I modified the summary components to watch the store
through a computed value.

Some of the error summaries contained the wrong focus and
scroll ids for the direct debit field. This should now be
fixed.

The summary was vanishing if the bankdata was invalid and
the rest of the form was valid, so I updated the
useAddressFormEventHandlers composable to mark it as invalid
properly and added a test.

I also noticed that the error summaries weren't displaying
all the errors on the alternate form when the donor had
selected yes for receipt and not selected an address type.

Ticket: https://phabricator.wikimedia.org/T374216