Make autocomplete autoscroll on focus mobile only
We added an on focus autoscroll to the autocomplete fields
to stop the dropdown menu being hidden behind the on screen
keyboard. This works well on mobile but is annoying on desktop,
so this makes the autoscroll only happen on smaller sizes.

I also moved the autoscroll functionality into a composable as
it's shared between 3 fields.

Ticket: https://phabricator.wikimedia.org/T374419