Make a prod ready prototype of the IBAN field
Adds an IBAN field that auto formats text on input.

Adds new BankDataSection that contains and validates
the IBAN field, and an IBAN calculator to allow donors
to find their IBAN using their bank account details.

Adds ButtonLink component to be used when the functionality
of a button is required, but it should look like a link.

Adds tests.

Ticket: https://phabricator.wikimedia.org/T374426