Add IBAN field to all versions of the donation form
We have 2 different versions of the donation form currently
and this adds the IBAN field to them.

Also moves the tests for the payment summaries out of the
payment summary tests and into the DonationForm component
tests because we now need to test the summaries with
fields not available in the personal data sections.

Ticket: https://phabricator.wikimedia.org/T374426