Add IBAN field to all versions of the donation form
We have 4 different versions of the donation form currently
and this adds the IBAN field to them.

Updated the donation form integration tests to check for the
new field.

Added and deprecated methods in the bankdata store in
preparation for simplifying it.

Hooked up the new IBAN only validation actions in the form
validation composable.

Ticket: https://phabricator.wikimedia.org/T374426