Fix payment field validation issue
The composable that handles the error summary and
form validation on the donation form was watching
a store getter that didn't exist.

This fixes that, and adds tests to make sure the
error summaries and form submits are handled on
all iterations of the donation form.

There was also a message about requiring an email
being displayed to anonymous donors, which is now
removed.