Silence SASS Deprecation Warnings
The version of Bulma we use now uses deprecated sass
functions. We can't update to the next version of Bulma
and we plan to eventually remove it altogether. In the
interim I've turned off deprecation warnings from the
sass compiler. We will eventually need to turn this back
on as our site will break when Dart Sass hits v3.

Ticket: https://phabricator.wikimedia.org/T378075