Lint TypeScript and Vue files again
Previously this was done with the deprecated `--ext` flag for ESLint
When I removed the flag, TypeScript and Vue files were no longer linted

Removed old fix for an eslint/typescript override problem that has been
fixed now. See https://stackoverflow.com/a/65768375/130121

Added a commented-out suggestion that needs to be discussed in the team:
Currently, we only use the `@typescript-eslint` plugin to be able to
parse TypeScript files. The TypeScript files are *not* linted with
TypeScript-specific rules from `@typescript-eslint`. I think this is an
oversight and we should do that in the future. For an easy start, we
should disable the 3 rules that give the most errors (most of them
stemming from legacy code)