Validate invalid text fields in input
We had reports from some donors that they get confused
when they fix a field that had been previously marked
invalid, because it stays invalid until they blur it.

Validating a field on every input is annoying to the
donors so we shouldn't do that.

This strikes a balance by making the validation check
run on input only when the field is invalid, and once
it becomes valid again then stops. It will then
validate again when they blur the field.

Ticket: https://phabricator.wikimedia.org/T390039