Create a pattern library
The tech team is taking ownership of our design system.
The first step in doing that is the creation of a
pattern library that provides a visual and technical
source of truth for all the parts of the site. This
is the beginning of that.

- Sets up basic pattern library structure
- Adds Cube CSS/Every Layout CSS compositions.
- Adds Visually Hidden utility.
- Adds patterns for:
	- Accordion
	- Button
	- Content Card
	- Donation Comment
	- Footer Nav
	- Icon Text
	- Locale Switcher
	- Mobile Nav Toggle
	- Modal Dialogue
	- More Info Toggle
	- Nav
	- Pagination
	- Section Heading
	- Site Foot
	- Site Head
	- Summary

Ticket: https://phabricator.wikimedia.org/T395100