Remove boolean types, use undefined instead
The two composables `useAddressSummary` and `useBankDataSummary`
returned both data and a boolean if data is available. This is weird
because you could express states that should be impossible (e.g.
`hasAddressSummary: false` with filled data and `hasAddressSummary:
true` with empty or undefined data). This commit uses a JavaScript
primitive for this: `undefined`. `undefined` is one of the [Falsy
values](https://developer.mozilla.org/en-US/docs/Glossary/Falsy) which
makes it nice for checks. We're using `undefined` instead of `null`
because we want to express the fact "there is absolutely no data here,
not even a null value". This distinguishes our type from nullable
optional types.

The type change also revealed a small bug in DonationSummary.spec.ts
where we accidentally expected empty strings, which could have made the
test pass where it should not pass. This commit also fixes that bug.

This commit also removes additional properties form the returned address
in `useAddressSummary` that would be ignored and are not used anywhere.