Fix margin on even 2 column radio
We remove the margin bottom of the last radio item
in order to keep the spacing consistent with the
next field group. This caused an issue when there
was an even number of items that caused the last
item to get bigger than it's siblings.

This fix makes sure all items in the final row
of a radio group have the margin bottom removed.

Ticket: https://phabricator.wikimedia.org/T397707