Add pattern library implementation scaffold
We want to start applying the pattern library into our
production code in a non-backwards breaking manner.
This means we first need to apply the HTML and classes
from the pattern library and style them so they match
the existing production styles.

This adds a scaffold to start the process:

- Imports the pattern library compositions in the production css
- Adds the CSS vars required by the compositions and set
  up the spacing to match production
- Adds a ContentCard and SectionHeading pattern components
- Hooks them up on the Membership Form, replacing FormSection
- Add pattern library compatibility folder and styles for the
  Content Card and Section Heading
- Renames the `.donation-page-form-section` class to `.content-card`
  and removes those specific styles from the donation form.
- Shuffles some misc css around to better places
- Deprecates FormSection. To deprecate this it needed to be
  converted to plain Composition API.

Ticket: https://phabricator.wikimedia.org/T399537