Add production implementation of icon text
This adds the icon text block/pattern to the production
site and some compatibility code.

Replaces all icon-title elements with the icon text pattern

It also adds a new extension to the icon text pattern
to allow headings to be styled as normal text. This is
useful for sidebars.

Ticket: https://phabricator.wikimedia.org/T396008