Remove the scroll target component
We have a fixed header which makes scrolling to an
element difficult as the scrolled to element will
be hidden under it. Previously we solved that with
hidden scroll elements that were moved up the page
with minus margin tops.

This removes all those and replaces is with a global
css scroll-margin-top rule.

Also adds tests to make sure all error summaries
have working scroll and focus targets.

Ticket: https://phabricator.wikimedia.org/T392206