Add flex field group composition
This is a new composition that allows us to
layout groups of fields in different ways.

Adds a sample of the compact form to the pattern
library.

Adds a variant to the field container for explicitly
specifying a max width field. Individual fields
should fill their containers, and we should use
compositions to control the layout.

Ticket: https://phabricator.wikimedia.org/T405438