Rename the data-max-width attribute
The field container has an exception that stops
it going over a certain width (data-max-width).

This is okay in the context of the field container,
but when implementing it we abstract it inside our
form fields. This leads to confusion because the
attribute name is so generic. To fix that this
commit changes the name of the attribute in the
containing fields and they now explicitly set the
data-max-width attribute on the field container.

Ticket: https://phabricator.wikimedia.org/T406561