Fix forms being submitted with enter key
Adds @submit.prevent to visible forms to prevent
them being submitted with the enter key.

Some of our forms had a <button> element in them
that was not marked as type=submit. This was
accidentally the thing that was blocking forms
from being submitted. This button was marked as
type=button in a recent change but some forms
didn't have their submit blocked elsewhere.